+++ elm/lib/dependabot/elm/version.rb
@@ -1,19 +1,21 @@
 # frozen_string_literal: true
 
+require "dependabot/utils"
+
 # Elm versions require major, minor and patch to be present
 # They don't allow any letters
 
 module Dependabot
-  module Utils
-    module Elm
-      class Version < Gem::Version
-        VERSION_PATTERN = "[0-9]+\.[0-9]+\.[0-9]+"
-        VERSION_PATTERN_REGEX = /\A#{VERSION_PATTERN}\Z/.freeze
+  module Elm
+    class Version < Gem::Version
+      VERSION_PATTERN = "[0-9]+\.[0-9]+\.[0-9]+"
+      VERSION_PATTERN_REGEX = /\A#{VERSION_PATTERN}\Z/.freeze
 
-        def self.correct?(version)
-          version.to_s.match?(VERSION_PATTERN_REGEX)
-        end
+      def self.correct?(version)
+        version.to_s.match?(VERSION_PATTERN_REGEX)
       end
     end
   end
 end
+
+Dependabot::Utils.register_version_class("elm_package", Dependabot::Elm::Version)
